Move functions out of entities to utils (#600)
