+++ explorer/app/apis/catalog/anvil-catalog/common/entities.ts
@@ -1,7 +1,6 @@
 /**
  * Model of AnVIL catalog.
  */
-import { sanitizeString } from "../../../../viewModelBuilders/common/utils";
 
 export type AnVILCatalogEntity =
   | AnVILCatalogStudy
@@ -46,46 +45,3 @@ export interface AnVILCatalogWorkspace {
   studyName: string;
   workspaceName: string;
 }
-
-/**
- * Returns the dbGapId.
- * @param workspaceOrStudy - AnVIL catalog entity.
- * @returns String value of dbGapId.
- */
-export const getDbGapId = (
-  workspaceOrStudy: Exclude<AnVILCatalogEntity, AnVILCatalogConsortium>
-): string => workspaceOrStudy.dbGapId || "";
-/**
- * Returns the terra workspace name.
- * @param anvilCatalogWorkspace - AnVIL catalog workspace.
- * @returns String value of terra workspace name.
- */
-export const getTerraWorkspaceName = (
-  anvilCatalogWorkspace: AnVILCatalogWorkspace
-): string => anvilCatalogWorkspace.workspaceName ?? "";
-/**
- * Returns the consortium.
- * @param anvilCatalogEntity - AnVIL catalog entity.
- * @returns String value of consortium.
- */
-export const getConsortium = (anvilCatalogEntity: AnVILCatalogEntity): string =>
-  anvilCatalogEntity.consortium ?? "";
-
-export const anvilCatalogStudyInputMapper = (
-  input: AnVILCatalogStudy
-): AnVILCatalogStudy => {
-  return {
-    ...input,
-    studyName: sanitizeString(input.studyName),
-  };
-};
-
-export const anvilCatalogWorkspaceInputMapper = (
-  input: AnVILCatalogWorkspace
-  // eslint-disable-next-line sonarjs/no-identical-functions -- //TODO remove this duplication
-): AnVILCatalogWorkspace => {
-  return {
-    ...input,
-    studyName: sanitizeString(input.studyName),
-  };
-};
