+++ explorer/app/viewModelBuilders/anvil-catalog/common/viewModelBuilders.ts
@@ -4,8 +4,8 @@ import {
   AnVILCatalogEntity,
   AnVILCatalogStudy,
   AnVILCatalogWorkspace,
-  getDbGapId,
 } from "../../../apis/catalog/anvil-catalog/common/entities";
+import { getStudyId } from "../../../apis/catalog/anvil-catalog/common/utils";
 import { ExploreState } from "../../../common/context/exploreState";
 import * as C from "../../../components";
 import {
@@ -228,7 +228,7 @@ export const buildStudyName = (
     links: [
       {
         label: workspaceOrStudy.studyName,
-        url: `/studies/${getDbGapId(workspaceOrStudy)}`,
+        url: `/studies/${getStudyId(workspaceOrStudy)}`,
       },
     ],
   };
